requirements.txt
streamlit
openai
requests
numpy
faiss-cpu
sentence-transformers
python-dotenv
reportlab